Fixing EventSource duplication and adding log collection script.
