+++ provisioning/device/src/Microsoft.Azure.Devices.Provisioning.Client.csproj
@@ -15,9 +15,6 @@
     <Compile Include="$(Common)\Logging.Common.cs">
       <Link>Common\Logging.Common.cs</Link>
     </Compile>
-    <Compile Include="$(Common)\device\provisioning\Logging.ProvisioningDeviceClient.cs">
-      <Link>Common\device\provisioning\Logging.ProvisioningDeviceClient.cs</Link>
-    </Compile>
   </ItemGroup>
 
   <ItemGroup>
