+++ provisioning/e2e/ProvisioningTests.cs
@@ -83,10 +83,9 @@ namespace Microsoft.Azure.Devices.E2ETests
                 Assert.IsNotNull(result.AssignedHub);
                 Assert.IsNotNull(result.DeviceId);
 
-                Client.IAuthenticationMethod auth = await CreateAuthenticationMethodFromSecurityProvider(
+                Client.IAuthenticationMethod auth = CreateAuthenticationMethodFromSecurityProvider(
                     security, 
-                    result.DeviceId, 
-                    result.AssignedHub).ConfigureAwait(false);
+                    result.DeviceId);
 
                 using (DeviceClient iotClient = DeviceClient.Create(result.AssignedHub, auth, Client.TransportType.Mqtt_Tcp_Only))
                 {
@@ -257,10 +256,9 @@ namespace Microsoft.Azure.Devices.E2ETests
             throw new NotSupportedException($"Unknown security type: '{name}'.");
         }
 
-        private async Task<Client.IAuthenticationMethod> CreateAuthenticationMethodFromSecurityProvider(
+        private Client.IAuthenticationMethod CreateAuthenticationMethodFromSecurityProvider(
             SecurityProvider provisioningSecurity,
-            string deviceId,
-            string iotHub)
+            string deviceId)
         {
             _verboseLog.WriteLine($"{nameof(CreateAuthenticationMethodFromSecurityProvider)}({deviceId})");
