+++ provisioning/transport/amqp/src/Logging.ProvisioningTransport.Amqp.cs
@@ -1,0 +1,13 @@
+// Copyright (c) Microsoft. All rights reserved.
+// Licensed under the MIT license. See LICENSE file in the project root for full license information.
+
+using System.Diagnostics.Tracing;
+
+namespace Microsoft.Azure.Devices.Shared
+{
+    // cc5b923d-ab24-57ee-bec8-d2f5cf1bb6e4
+    [EventSource(Name = "Microsoft-Azure-Devices-Provisioning-Transport-Amqp")]
+    internal sealed partial class Logging : EventSource
+    {
+    }
+}
