+++ provisioning/transport/http/src/Microsoft.Azure.Devices.Provisioning.Transport.Http.csproj
@@ -17,9 +17,9 @@
     <Compile Include="$(Common)\Logging.Common.cs">
       <Link>Common\Logging.Common.cs</Link>
     </Compile>
-    <Compile Include="$(Common)\device\provisioning\Logging.ProvisioningDeviceClient.cs">
-      <Link>Common\device\provisioning\Logging.ProvisioningDeviceClient.cs</Link>
-    </Compile>
+    <Compile Include="$(Common)\Logging.ProvisioningTransport.Common.cs">
+      <Link>Common\Logging.ProvisioningTransport.Common.cs</Link>
+    </Compile>  
     <Compile Include="$(Common)\device\provisioning\transport\ClientApiVersionHelper.cs"> 
       <Link>Common\device\provisioning\transport\ClientApiVersionHelper.cs</Link>
     </Compile>
