+++ tools/CaptureLogs/iot_providers.txt
@@ -1,0 +1,5 @@
+"{e927240b-7198-5cc8-72f1-7ddcf31bb8cb}"
+"{cc5b923d-ab24-57ee-bec8-d2f5cf1bb6e4}"
+"{d209b8a1-2e02-5724-f341-677227d0ed22}"
+"{2143dadd-f500-5ff9-12b3-9afacae4d54c}"
+"{d079e771-0495-4124-bd2f-ab63c2b50525}"
