+++ tools/CaptureLogs/iot_startlog.cmd
@@ -1,0 +1,2 @@
+logman create trace IotTrace -o iot.etl -pf iot_providers.txt
+logman start IotTrace
