+++ tools/CaptureLogs/iot_stoplog.cmd
@@ -1,0 +1,2 @@
+logman stop IotTrace
+logman delete IotTrace
